fix: Unify reserveCount and maxReserve to uint16

+++ contracts/mocks/helpers/MockPool.sol
@@ -31,7 +31,7 @@ contract MockPool {
 import {Pool} from '../../protocol/pool/Pool.sol';
 
 contract MockPoolInherited is Pool {
-  uint256 internal _maxNumberOfReserves = 128;
+  uint16 internal _maxNumberOfReserves = 128;
 
   function getRevision() internal pure override returns (uint256) {
     return 0x3;
@@ -39,7 +39,7 @@ contract MockPoolInherited is Pool {
 
   constructor(IPoolAddressesProvider provider) Pool(provider) {}
 
-  function setMaxNumberOfReserves(uint256 newMaxNumberOfReserves) public {
+  function setMaxNumberOfReserves(uint16 newMaxNumberOfReserves) public {
     _maxNumberOfReserves = newMaxNumberOfReserves;
   }
